Moving Git Issues to seperate tab and reorodering tabs
