+++ src/ApiDiscussionsResults.js
@@ -11,16 +11,7 @@ const discussionEndpoints = [
   'https://epic-health-dashboard-artifacts.s3.amazonaws.com/discussions-ufs-community-ufs-weather-model.json',
 ];
 
-const issueEndpoints =  [
-  'https://epic-health-dashboard-artifacts.s3.amazonaws.com/issues-hafs-community-HAFS.json',
-  'https://epic-health-dashboard-artifacts.s3.amazonaws.com/issues-NOAA-EMC-UPP.json',
-  'https://epic-health-dashboard-artifacts.s3.amazonaws.com/issues-NOAA-EPIC-land-offline_workflow.json',
-  'https://epic-health-dashboard-artifacts.s3.amazonaws.com/issues-ufs-community-land-DA_workflow.json',
-  'https://epic-health-dashboard-artifacts.s3.amazonaws.com/issues-ufs-community-ufs-srweather-app.json',
-  'https://epic-health-dashboard-artifacts.s3.amazonaws.com/issues-ufs-community-ufs-weather-model.json',
-];
-
-function ApiTestResults() {
+function ApiDiscussionResults() {
   return (
     <div style={{ padding: 30 }}>
       <div>
@@ -30,14 +21,8 @@ function ApiTestResults() {
         <p><a href="https://github.com/orgs/ufs-community/discussions">UFS Community Discussions</a></p>
         <ItemDataGrid endpoints={discussionEndpoints} />
       </div>
-      <div>
-        <h1>Issues</h1>
-        <p>Includes: ufs-srweather-app, ufs-weather-model, UPP, land-DA_workflow, HAFS, and NOAA-EPIC/land-offline_workflow</p>
-        <p><span style={{ color: 'red' }}>Red</span> indicates that there has been no initial reply to the author.</p>
-        <ItemDataGrid endpoints={issueEndpoints} />
-      </div>
     </div>
   );
 }
 
-export default ApiTestResults;
+export default ApiDiscussionResults;
